Move accountContext in /core
