+++ packages/core/src/tools/abstractHooks.ts
@@ -26,6 +26,8 @@ export const SKIPPED = 'SKIPPED'
 export type DISCARDED = typeof DISCARDED
 export type SKIPPED = typeof SKIPPED
 
+export type AbstractHooks = ReturnType<typeof abstractHooks>
+
 export function abstractHooks<T extends { [K in HookNames]: (...args: any[]) => any }, E>() {
   const callbacks: { [K in HookNames]?: Array<T[K]> } = {}
