+++ packages/core/test/index.ts
@@ -26,3 +26,4 @@ export * from './disableJasmineUncaughtExceptionTracking'
 export * from './instrumentation'
 export * from './wait'
 export * from './consoleLog'
+export * from './createHooks'
